#184 Review PR changes

+++ src/main/math/Math.hpp
@@ -611,7 +611,7 @@ namespace nta {
   template <>
   struct Sqrt<long double> : public std::unary_function<long double,long double>
   {
-    inline long double operator()(const long double& x) const { return sqrtl(x); }
+    inline long double operator()(const long double& x) const { return sqrt(x); }
   };
 
   template <typename T>
@@ -639,7 +639,7 @@ namespace nta {
   template <>
   struct Exp<long double> : public std::unary_function<long double,long double>
   {
-    inline long double operator()(const long double& x) const { return expl(x); }
+    inline long double operator()(const long double& x) const { return exp(x); }
   };
 
   template <typename T>
@@ -661,7 +661,7 @@ namespace nta {
   template <>
   struct Log<long double> : public std::unary_function<long double,long double>
   {
-    inline long double operator()(const long double& x) const { return logl(x); }
+    inline long double operator()(const long double& x) const { return log(x); }
   };
 
   template <typename T>
@@ -702,7 +702,7 @@ namespace nta {
 #if defined(NTA_PLATFORM_win32) || defined(NTA_PLATFORM_win64)
       return log(x) / log(2.0);
 #else
-      return log2l(x); 
+      return log2(x); 
 #endif
     }
   };
@@ -745,7 +745,7 @@ namespace nta {
 #if defined(NTA_PLATFORM_win32) || defined(NTA_PLATFORM_win64)
       return log(x) / log(10.0);
 #else
-      return log10l(x); 
+      return log10(x); 
 #endif
     }
   };
@@ -788,7 +788,7 @@ namespace nta {
 #if defined(NTA_PLATFORM_win32) || defined(NTA_PLATFORM_win64)
       return log(1.0 + x);
 #else
-      return log1pl(x); 
+      return log1p(x); 
 #endif
     }
   };
@@ -875,7 +875,7 @@ namespace nta {
   {
     inline long double operator()(const long double& x, const long double& y) const
     { 
-      return powl(x,y); 
+      return pow(x,y); 
     }
   };
 
@@ -907,7 +907,7 @@ namespace nta {
   {
     inline long double operator()(const long double& x, const long double& y) const
     { 
-      return logl(x)/logl(y); 
+      return log(x)/log(y); 
     }
   };
